refactor : step3 레이싱 정책 클래스 추가
